feat(swagger): Provide URI version to operationIdFactory
Adds a new optional parameter to operationIdFactory to allow providing the API version prefix added to the operation's path.

This allows the operationIdFactory to generate unique IDs in the case of the same controller method being used for multiple API versions.

Default behavior is unchanged - the operation IDs will be of the form controllerKey_methodKey, but where necessary this can be customised using the URI version prefix.

This is based on the work completed in https://github.com/nestjs/swagger/pull/1949. Those changes were reverted due to test failures.

This change adapts those changes to also support neutral versions which resolves the test failures.

Closes nestjs#2537